! break managefeatures controller into bite size chunks
